implement splice-in (step 1: create)
We closely mimick rbf-ing the initial funding tx (`RbfStatus` vs `SpliceStatus`, reuse of `TxInitRbf`/`TxAckRbf` messages).

Thre main design considerations:
- we use a poor man's _quiescence_ protocol which just reject the splice if the channel is not idle
- splice txs always _spend_ the previous funding/splice tx, even if it isn't confirmed yet and could theoretically be RBFed. This is done to be compatible with zero-conf splices, as we intend to use this for Phoenix
- in `InteractiveTxBuilder`, we introduce a new "common input" (`Multisig2of2Input`), for the previous funding outpoint. It is not exchanged between peers, each side adds it implicitly in case of a splice

A new `SpliceTx` purpose is introduced, which is essentially the same as `FundingTxRbf` except that we start from the previous commitment index.

Only the initiator contribute to the splice (but the wiring for two-ways splices is already there).