implement splice-out
Splice-out may seem simpler than splice-in, but it comes with its own challenges:
- in splice-in, we always take the previous funding output, and add more to it
- in splice-out, in order to pay for the splice tx fees, we may need to take from the existing funding output. There is a chicken-and-egg issue due to the fact that, in the general case, we don't know the size of the splice tx until this end of the interactive tx process.

To deal with this, I assumed that we don't mix splice-in and splice-out (although everything is wired to allow it).

Note that, in case of a dual splice-out (even if it is not available from the API), we currently have the initiator pay for both, and use a hardcoded weight that will be completely inaccurate.